feat(final-adjustments): added new project
