+++ src/util-constants/index.ts
@@ -1,1 +1,2 @@
 export * from './personal-information';
+export * from './projects';
